pmcf/resources/aws/cloudformation.py
# Copyright (c) 2014 Piksel
#
#    Licensed under the Apache License, Version 2.0 (the "License"); you may
#    not use this file except in compliance with the License. You may obtain
#    a copy of the License at
#
#         http://www.apache.org/licenses/LICENSE-2.0
#
#    Unless required by applicable law or agreed to in writing, software
#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
#    License for the specific language governing permissions and limitations
#    under the License.

from troposphere import cloudformation as cfn

from pmcf.resources.aws.helpers import cloudformation as cf
from pmcf.utils import error


class Stack(cf.Stack):
    def JSONrepr(self):
        try:
            return super(self.__class__, self).JSONrepr()
        except ValueError, e:
            error(self, e.message)


class WaitCondition(cfn.WaitCondition):
    def JSONrepr(self):
        try:
            return super(self.__class__, self).JSONrepr()
        except ValueError, e:
            error(self, e.message)


class WaitConditionHandle(cfn.WaitConditionHandle):
    pass


__all__ = [
    Stack,
    WaitCondition,
    WaitConditionHandle,
]
